Add support for parameters and recorders stored as node attributes.
This change allows parameter and recorder datasets to be stored as nodal attributes in hydra. When exporting to the Pywr namespace these attributes are convereted to the appropriate (parameter or recorder) global namespace by prefixing with the node's name. The format is is "__{node_name}__:{attribute_name}". The importer also recognises this format and will import such components as nodal attributes.